bugfixing: deleted time-period.js because it was rerendering the time-period
