--- pages/TripDetailPage/[slug].js
@@ -1,13 +1,0 @@
-import { useRouter } from "next/router";
-
-export default function CategoryPage() {
-  const router = useRouter();
-  const { categorySlug, pageNumber } = router.query;
-
-  return (
-    <div>
-      <p>Category Slug: {categorySlug}</p>
-      <p>Page Number: {pageNumber}</p>
-    </div>
-  );
-}
